Instead of hiding the console window, free the console
This allows us to pass our console on to the child process, then free it
so we don't have to deal with it anymore. Now if you start a process
like pwsh directly it won't just immediately hide its console. Also, GUI
(Windows subsystem) apps that don't use a console will just let it be
destroyed.

We also now don't need a CTRL-C handler, since that only affects apps
with consoles.