Use job objects to manage child processes
There was an issue with Java apps like
https://github.com/hneemann/Digital/ where the JVM would say it couldn't
read the jar file. At first I thought this was do to weird attributes on
the virtualized files. Then I realized the issue was that Digital.exe
was launching the JVM and immediately exiting. This meant that the
loader's call to wait would return and the provider would be shut down,
even though a child process of the process we launched was still
running.

It seems the best way to manage process trees in Windows is with job
objects. I didn't put any resource limits on the job. I simply created
the child in a job so that all of its children would also be in the job,
and then waited for all processes in the job to exit.